users dns

+++ Shell/manage_fileDNS.sh
@@ -15,9 +15,17 @@ function addline_fileDNS() {
 		echo "+$1.myshop.itinet.fr:88.177.168.133:1800" >> /etc/tinydns/root/data
 		echo "Ajout [OK]"
 		
-		rm /etc/tinydns/root/data.cdb
-		/usr/local/bin/tinydns-data
-		
+		echo "Mise à jour du fichier de zone en cours..."
+		if -e /etc/tinydns/root/data.cdb; then
+			rm /etc/tinydns/root/data.cdb
+			cd /etc/tinydns/root
+			/usr/local/bin/tinydns-data
+		echo "Mise à jour [OK]"
+		else
+			echo "Le fichier data.cdb n'existe pas dans /etc/tinydns/root"
+			echo "Mise à jour [ECHEC]"
+		fi
+
 		ssh root@dedibox.itinet.fr
 	fi
 }
@@ -35,8 +43,16 @@ function manage_fileDNS {
 			echo "Renommage de $2 en $3 en cours..."
 			sed -i -e "s/$2/$3/g" /etc/tinydns/root/data
 			echo "Renommage [OK]"
-			rm /etc/tinydns/root/data.cdb 
-			/usr/local/bin/tinydns-data
+			echo "Mise à jour du fichier de zone en cours..."
+			if -e /etc/tinydns/root/data.cdb; then
+				rm /etc/tinydns/root/data.cdb 
+				/usr/local/bin/tinydns-data
+				echo "Mise à jour [OK]"
+			else
+				echo "Le fichier data.cdb n'existe pas dans /etc/tinydns/root"
+				echo "Mise à jour [ECHEC]"
+			fi
+
 			ssh root@dedibox.itinet.fr
 		else
 			echo "Cet enregistrement n'existe pas ! [ECHEC]"
@@ -49,8 +65,16 @@ function manage_fileDNS {
 			cp /etc/tinydns/root/data2 /etc/tinydns/root/data
 			rm /etc/tinydns/root/data2
 			echo "Suppression [OK]"
-			rm /etc/tinydns/root/data.cdb
-			/usr/local/bin/tinydns-data
+			echo "Mise à jour du fichier de zone en cours..."
+			if -e /etc/tinydns/root/data.cdb; then
+				rm /etc/tinydns/root/data.cdb
+				/usr/local/bin/tinydns-data
+				echo "Mise à jour [OK]"
+			else
+				echo "Le fichier data.cdb n'existe pas dans /etc/tinydns/root"
+				echo "Mise à jour [ECHEC]"
+			fi
+			
 			ssh root@dedibox.itinet.fr
 		else
 			echo "Cet enregistrement n'existe pas ! [ECHEC]"
